refactor assets for POC. add test cases, refactor out unused files (#65)
* refactor assets for POC. add test cases, refactor out unused files
* refactor to get code coverage to 100, lint errors to 2
* ignore fetch for lint to 1
* moving image to scss, import common css in application.scss
