+++ .eslintrc
@@ -7,10 +7,12 @@
 ],
 "globals": {
   "document": true,
-  "window": true
+  "window": true,
+  "fetch": false
   },
 
   "rules": {
+      "no-unused-vars": [0, { "vars": 'local', "args": "after-used", "ignoreRestSiblings": false }],
     "react/require-extension": "off",
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }] }
 }
