+++ app/assets/stylesheets/application.scss
@@ -1,6 +1,6 @@
 @import "bootstrap-sprockets";
 @import "bootstrap";
-@import "facility/facility.scss";
+@import "shared/headers.scss";
 /*
  * This is a manifest file that'll be compiled into application.css, which will include all the files
  * listed below.
