--- app/assets/stylesheets/facility/facility.scss
@@ -1,151 +1,0 @@
-body {
-  margin: 0 auto;
-  padding: 0;
-}
-/***********Stylesheet for Facility Profil Page***************/
-.facility-details {
-  padding: 0px;
-}
-.facility-details .facility_header, .facility-details .facility_header *{
-  padding: 0px;
-}
-.main_page .header_cwds {
-  background: #363636;
-  padding: 5px 20px;
-}
-.facility-details .header_left {
-  float: left;
-  margin-bottom: 0px;
-  font-size: 28px;
-  color: #8b8b8b;
-  font-family: "Source Sans Pro Bold", sans-serif;
-}
-.facility-details .header_right {
-  display: inline;
-  float: right;
-  margin-bottom: 0px;
-  margin-top: 33px;
-}
-.header_right .header_bold {
-  margin-bottom: 0px;
-  font-family: "Source Sans Pro Bold", sans-serif;
-  font-size: 14px;
-  display: inline;
-  color: #8d8d8d;
-}
-.header_right .header_light {
-  margin-bottom: 0px;
-  font-family: "Source Sans Pro ExtraLight", sans-serif;
-  color: #2b2b2b;
-  font-size: 16px;
-  display: inline;
-  font-weight: 100;
-}
-.facility-list .list_block {
-  padding: 0px 20px;
-  border: 1px solid #337ab7;
-  border-radius: 5px;
-  margin-bottom: 20px;
-}
-.facility-list .list_block h1 {
-  font-size: 28px;
-}
-.facility-list .list_block > div {
-  padding: 0px;
-}
-.facility-address .map_image .map{
-  height: 188px;
-}
-.facility-details-block {
-  padding: 20px 0px;
-}
-.facility-details-block > div {
-  padding: 0px;
-  padding-right: 138px;
-}
-.headliner {
-  margin-bottom: 30px;
-  border-bottom: 1px solid;
-  padding: 0px;
-}
-.headliner div{
-  width: 30%;
-  height: 2px;
-  background: #000;
-}
-.facility-details .facility_blocks {
-  padding: 0px;
-}
-.facility-details .inner_block{
-  height: 212px;
-  padding: 20px 10px 20px 20px;
-}
-.facility-address .inner_block {
-  padding: 20px 10px 20px 5px;
-}
-.facility-address .inner_block:last-child {
-  padding-left: 20px;
-}
-.facility-details .inner_block {
-  background-color: #f5f5f5;
-  border-left: 10px solid #8f8f8f;
-}
-//.facility-details .inner_block.yellow {
-//  border-left: 10px solid yellow;
-//}
-//.facility-details .inner_block.green {
-//  border-left: 10px solid green;
-//}
-//.facility-details .inner_block.orange {
-//  border-left: 10px solid darkorange;
-//}
-.small_inner_block p:first-child {
-  margin-bottom: 0px;
-  color: #8d8d8d;
-  font-size: 12px;
-  font-family: "Source Sans Pro Bold", sans-serif;
-  font-weight: bold;
-}
-.small_inner_block p:last-child {
-  margin-bottom: 30px;
-  color: #2b2b2b;
-  font-size: 16px;
-  font-family: "Source Sans Pro ExtraLight", sans-serif;
-  font-weight: 100;
-}
-.facility-address .facility-address-block{
-  background-color: #f5f5f5;
-}
-.facility-address .facility-address-block .facility-address-details{
-  padding: 0px;
-}
-.physical_address, .postal_address, .county_details {
-  margin-bottom: 30px;
-}
-.visit_reason, .last_visit {
-  padding-left: 20px;
-}
-.facility-address .address_street {
-  margin-bottom: 0px;
-  color: #2b2b2b;
-  font-size: 16px;
-  font-family: "Source Sans Pro ExtraLight", sans-serif;
-  font-weight: 100;
-}
-.phone_primary {
-  display: inline;
-  float: left;
-  margin-right: 20px;
-}
-.phone_alt {
-  display: inline-block;
-
-}
-.phone_details .small_inner_block p:last-child {
-  margin-bottom: 0px;
-}
-.facility-address .last_visit .address_street {
-  margin-bottom: 52px;
-}
-
-@import "mobile/facility_mobile.scss";
