+++ app/assets/stylesheets/shared/headers.scss
@@ -1,0 +1,44 @@
+.header_right .header_bold
+{
+    font-family: 'Source Sans Pro Bold', sans-serif;
+    font-size: 14px;
+
+    display: inline;
+
+    margin-bottom: 0;
+
+    color: #8d8d8d;
+}
+.header_right .header_light
+{
+    font-family: 'Source Sans Pro ExtraLight', sans-serif;
+    font-size: 16px;
+    font-weight: 100;
+
+    display: inline;
+
+    margin-bottom: 0;
+
+    color: #2b2b2b;
+}
+.main_page .header_cwds
+{
+    /*@extend .bg-success;*/
+    padding: 5px 20px;
+
+    background: #363636;
+}
+body
+{
+    margin: 0 auto;
+    padding: 0;
+}
+
+.header-logo
+{
+    background-image: url('logo.jpg');
+    border: 0;
+    border-radius: 0;
+    height: 60px;
+    width: 150px;
+}
