+++ app/javascript/facilities/stylesheets/facilities.scss
@@ -1,0 +1,16 @@
+.facility-list .list_block
+{
+    margin-bottom: 20px;
+    padding: 0 20px;
+
+    border: 1px solid #337ab7;
+    border-radius: 5px;
+}
+.facility-list .list_block h1
+{
+    font-size: 28px;
+}
+.facility-list .list_block > div
+{
+    padding: 0;
+}
