+++ app/javascript/facility/stylesheets/facility.scss
@@ -1,9 +1,3 @@
-body
-{
-    margin: 0 auto;
-    padding: 0;
-}
-/***********Stylesheet for Facility Profil Page***************/
 .facility-details
 {
     padding: 0;
@@ -12,12 +6,6 @@ body
 .facility-details .facility_header *
 {
     padding: 0;
-}
-.main_page .header_cwds
-{
-    padding: 5px 20px;
-
-    background: #363636;
 }
 .facility-details .header_left
 {
@@ -38,41 +26,7 @@ body
     margin-top: 33px;
     margin-bottom: 0;
 }
-.header_right .header_bold
-{
-    font-family: 'Source Sans Pro Bold', sans-serif;
-    font-size: 14px;
 
-    display: inline;
-
-    margin-bottom: 0;
-
-    color: #8d8d8d;
-}
-.header_right .header_light
-{
-    font-family: 'Source Sans Pro ExtraLight', sans-serif;
-    font-size: 16px;
-    font-weight: 100;
-
-    display: inline;
-
-    margin-bottom: 0;
-
-    color: #2b2b2b;
-}
-.facility-list .list_block
-{
-    margin-bottom: 20px;
-    padding: 0 20px;
-
-    border: 1px solid #337ab7;
-    border-radius: 5px;
-}
-.facility-list .list_block h1
-{
-    font-size: 28px;
-}
 .facility-address .map_image .map
 {
     height: 188px;
@@ -291,10 +245,28 @@ body
         margin-bottom: 0;
     }
 }
-@media only screen and (min-width: 768px) and (max-width: 840px)
-{
-    .facility-details .inner_block
-    {
-        height: 263px;
-    }
+@media only screen and (max-width: 787px) {
+  .facility_blocks > div {
+    margin-bottom: 20px;
+  }
+  .facility-address .inner_block:last-child {
+    padding: 20px 20px 20px 15px;
+  }
+  .visit_reason, .last_visit {
+    padding-left: 0px;
+  }
+  .facility-address-block > div {
+    padding-left: 0px;
+  }
+  .postal_address {
+    margin-bottom: 0px;
+  }
+  .visit_reason p:last-child {
+    margin-bottom: 0px;
+  }
+}
+@media only screen and (min-width: 768px) and (max-width: 840px) {
+  .facility-details .inner_block{
+    height: 263px;
+  }
 }
