--- app/javascript/packs/hello.js
@@ -1,4 +1,0 @@
-import { renderReact } from 'hypernova-react'
-import hello from '../hello'
-
-export default renderReact('hello', hello)
