+++ app/javascript/search/search_input.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { FacilityTypes, CountyList } from '../constants/constants'
-import search_off from './images/search_off.png'
+import SearchOff from './images/search_off.png'
 
 export default class SearchInput extends React.Component {
   constructor (props) {
@@ -25,7 +25,7 @@ export default class SearchInput extends React.Component {
   }
   sendSearchInput (event) {
     event.preventDefault()
-    const finalQuery = (this.state.county + ',' + this.state.facility + ',' + (this.refs.facilityHome.value.length == 0 ? this.refs.facilityId.value : this.refs.facilityId.value + ',' + this.refs.facilityHome.value))
+    const finalQuery = (this.state.county + ',' + this.state.facility + ',' + (this.refs.facilityHome.value.length === 0 ? this.refs.facilityId.value : this.refs.facilityId.value + ',' + this.refs.facilityHome.value))
     this.props.sendSearchInput(finalQuery)
   }
   render () {
@@ -84,7 +84,7 @@ export default class SearchInput extends React.Component {
           </div>
           <div className='field_search col-xs-12 col-sm-1 col-md-1 col-lg-1'>
             <div className='search_block'>
-              <button type='submit'><img src={search_off} /></button>
+              <button type='submit'><img src={SearchOff} /></button>
             </div>
           </div>
         </form>
