+++ app/views/facilities/index.html.erb
@@ -1,5 +1,5 @@
 <div id = 'root'>
     <%= render_react_component('facilities', facilities: @facilities) %>
     <%= javascript_pack_tag 'facilities' %>
-    <%= stylesheet_pack_tag 'facility' %>
+    <%= stylesheet_pack_tag 'facilities' %>
 </div>
