--- app/views/pages/index.html.erb
@@ -1,4 +1,0 @@
-<div id = 'test'>
-    <%=render_react_component('hello', liked: @liked, title: @title) %>
-    <%= javascript_pack_tag 'hello'%>
-</div>
