+++ spec/cassettes/facilities_controller/get_search_renders_search.yml
@@ -1,0 +1,46 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: http://localhost:3001/v1/facilities/search?query=01%2C%2C%2Chood
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - Faraday v0.9.2
+  response:
+    status:
+      code: 200
+      message: 
+    headers:
+      x-frame-options:
+      - SAMEORIGIN
+      x-xss-protection:
+      - 1; mode=block
+      x-content-type-options:
+      - nosniff
+      content-type:
+      - application/json; charset=utf-8
+      etag:
+      - W/"afc31319bde362496769054005de46a8"
+      cache-control:
+      - max-age=0, private, must-revalidate
+      x-request-id:
+      - f35c1ca6-dee7-422e-93d5-b44f8103ad5f
+      x-runtime:
+      - '0.034983'
+      transfer-encoding:
+      - chunked
+    body:
+      encoding: UTF-8
+      string: '[{"fac_name":"ROBIN HOOD SFH","fac_nbr":193600029,"type":"Small Family
+        Home","fac_licensee_name":"HOOD, ROBIN AND MARIAN","fac_mail_street_addr":"7382
+        W. GRANDROSE AVE.","fac_mail_city":"LOS ANGELES","fac_mail_state":"CA","fac_mail_zip_code":"90063","fac_res_street_addr":"7389
+        W. GRANDROSE AVE.","fac_res_city":"LOS ANGELES","fac_res_state":"CA","fac_res_zip_code":"90063","county":"Los
+        Angeles","facility_telephone":"9098901041","district_office":"LA METRO \u0026
+        VALLEY CR","status":"Licensed","assigned_worker":"Kari Gutierrez","fac_licensee_type":"B","fac_lic_eff_date":"2016-08-30","fac_last_visit_date":"2016-09-03","last_visit_reason":"Renewal
+        (Fac.''s w/Expir.)","fac_email_address":null,"fac_capacity":6,"fac_orig_appl_rec_date":null}]'
+    http_version: 
+  recorded_at: Wed, 17 May 2017 22:19:39 GMT
+recorded_with: VCR 3.0.3
